+++ src/channel.ts
@@ -1,5 +1,5 @@
 import { ChannelState } from './channel_state';
-import { logChatPromiseExecution, messageSetPagination, normalizeQuerySort } from './utils';
+import { generateChannelTempCid, logChatPromiseExecution, messageSetPagination, normalizeQuerySort } from './utils';
 import { StreamChat } from './client';
 import {
   APIResponse,
@@ -1179,13 +1179,12 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
       this.cid = state.channel.cid;
       // set the channel as active...
 
-      const membersStr = state.members
-        .map((member) => member.user_id || member.user?.id)
-        .sort()
-        .join(',');
-      const tempChannelCid = `${this.type}:!members-${membersStr}`;
+      const tempChannelCid = generateChannelTempCid(
+        this.type,
+        state.members.map((member) => member.user_id || member.user?.id || ''),
+      );
 
-      if (tempChannelCid in this.getClient().activeChannels) {
+      if (tempChannelCid && tempChannelCid in this.getClient().activeChannels) {
         // This gets set in `client.channel()` function, when channel is created
         // using members, not id.
         delete this.getClient().activeChannels[tempChannelCid];
@@ -1211,7 +1210,7 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
       }),
     };
 
-    this.getClient().polls.hydratePollCache(messageSet.messages, true);
+    this.getClient().polls.hydratePollCache(state.messages, true);
 
     const areCapabilitiesChanged =
       [...(state.channel.own_capabilities || [])].sort().join() !==