+++ test/unit/connection.js
@@ -77,6 +77,14 @@ describe('connection', function () {
 			expect(data.device).to.deep.equal(device);
 		});
 
+		it('should properly encode X-Stream-Client', () => {
+			const userAgent = 'agent|val=foo bar';
+			client.userAgent = userAgent;
+			const url = ws._buildUrl();
+
+			expect(url).to.contain(encodeURIComponent(userAgent));
+		});
+
 		it('should not include device if not there', function () {
 			ws.client.options.device = undefined;
 			const { query } = url.parse(ws._buildUrl(), true);